Try separate searchfield as control
